Make structure for web development
I built out the basics of the this file structure according to the
doumentation found in learn. I've set up nodemon to run the dev
environment at loalhost:1234 via Express. This will currently serve up
the `client/dist` folder, which is why the test file is in there.

I'm including the package-lock so we can ensure we're all using the
correct versions of the dependencies, which I'm assuming we should make
routine for the project.