Implement dotenv for environment variables
For making API calls, we will need to be using our Github token, which
should be secure and not checked into the repo. dotenv lets you store
whatever in a `.env` file that can be recalled with `process.env.NAME`,
with NAME being the key of a `KEY=VALUE` pair within the `.env` file.